вроде есть простая логика чтения

+++ src/worker/WorkerManager.cpp
@@ -5,7 +5,7 @@
 using namespace mi::worker;
 
 WorkerManager::WorkerManager(int workersCount) : _workers() {
-    for (size_t i = 0; i < workersCount; i++) {
+    for (int i = 0; i < workersCount; i++) {
         _workers.emplace_back(Worker{i});
     }
